@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.10.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -31,8 +31,6 @@ dependencies:
 
 #  provider: ^4.1.3
   states_rebuilder: ^4.3.0
-#  meta: ^1.1.8
-#  equatable: ^1.2.0
   firebase_auth: ^3.0.1
   firebase_core: ^1.4.0
   cloud_firestore: ^2.4.0
@@ -40,7 +38,6 @@ dependencies:
   path: ^1.8.0
   intl: ^0.17.0
   logger: ^1.0.0
-  http: ^0.13.3
   file_picker: ^3.0.4
   dio: ^4.0.0
 
