@@ -1,10 +1,10 @@
 class AuthorizationException extends Error {
-  final String message;
+  final String? message;
 
   AuthorizationException(this.message);
 
   @override
   String toString() {
-    return message;
+    return message!;
   }
 }
