@@ -1,7 +1,4 @@
 import 'package:flutter/foundation.dart';
-import 'package:iothub/src/domain/entities/device.dart';
 
 @immutable
-class DeviceMeasurement {
-
-}
+class DeviceMeasurement {}
