@@ -16,7 +16,7 @@ class IOTHubService {
     return await _dbRepository.loadAllIOTHubs();
   }
 
-  Future<List<Device>> loadAllDevices(String selectedIOTHubID) async {
+  Future<List<Device>> loadAllDevices(String? selectedIOTHubID) async {
     if (selectedIOTHubID == null || selectedIOTHubID.isEmpty) {
       throw DatabaseException('NO IOT HUb selected');
     }
@@ -24,7 +24,7 @@ class IOTHubService {
   }
 
   Future<List<Measurement>> loadLastMeasurement(
-      String iothubId, Device device) async {
+      String? iothubId, Device device) async {
     return await _dbRepository.loadLastMeasurement(iothubId, device);
   }
 
